+++ scripts/details-popup/details-popup.js
@@ -34,7 +34,7 @@ const updateDetailsPopup = (detailsPopup, id) => {
 }
 
 (() => {
-	chrome.storage.local.get(["settings", "kana_vocab_assoc", "kanji_assoc", "vocabulary_assoc"], result => {
+	chrome.storage.local.get(["settings", "kana_vocab_assoc", "kanji_assoc", "vocabulary_assoc", "highlight_setup", "lessons", "reviews", "radical_id_list"], result => {
 		chrome.runtime.sendMessage({uptimeDetailsPopup:true});
 
 		const settings = result["settings"];
@@ -78,34 +78,12 @@ const updateDetailsPopup = (detailsPopup, id) => {
 			if (request.infoPopupFromSearch && !atWanikani)  {
 				console.log(request.infoPopupFromSearch, atWanikani, result);
 				let id = request.infoPopupFromSearch;
-				if (id.includes("rand")) {
-					let assocs = {};
-					if (id == "rand") {
-						assocs = {...result["kana_vocab_assoc"], ...result["kanji_assoc"], ...result["vocabulary_assoc"]};
-						console.log(assocs);
-					}
-					else if (id == "rand-kanji") {
-						assocs = result["kanji_assoc"];
-					}
-					else if (id == "rand-vocabulary") {
-						assocs = {...result["vocabulary_assoc"], ...result["kana_vocab_assoc"]};
-					}
-					const ids = Object.values(assocs);
-					// get random id
-					id = ids[Math.floor(Math.random() * ids.length)];
-				}
+
+				if (id.includes("rand"))
+					id = subjectRandomId(id, result);	
 				
 				id = Number(id);
 
-				// TODO: random kanji/vocab
-				/*if (id.split("-")[0] === "rand") {
-					let allSubjectsKeys = [Object.keys(allKanji), Object.keys(allVocab)].flat(1);
-					if (id.split("-")[1] === "kanji") allSubjectsKeys = Object.keys(allKanji);
-					else if (id.split("-")[1] === "vocabulary") allSubjectsKeys = Object.keys(allVocab);
-
-					if (allSubjectsKeys)
-						id = allSubjectsKeys[rand(0, allSubjectsKeys.length-1)];
-				}*/
 				console.log(id, detailsPopup);
 				if (!detailsPopup)
 					detailsPopup = await createDetailsPopup(id, settings);